@@ -20,9 +20,9 @@
 	"prefer-stable": true,
 	"prefer-dist": true,
 	"require": {
-		"php": ">=5.6",
-		"icanboogie/icanboogie": "^4.0",
-		"icanboogie/render": "^0.6"
+		"php": "^7.2",
+		"icanboogie/icanboogie": "^5.0",
+		"icanboogie/render": "^0.7"
 	},
 	"autoload": {
 		"psr-4": {
