@@ -11,13 +11,13 @@
 
 namespace ICanBoogie\Binding\Render;
 
-use function ICanBoogie\app;
-use function ICanBoogie\get_autoconfig;
 use ICanBoogie\Application;
 use ICanBoogie\Render;
 use ICanBoogie\Render\EngineCollection;
 use ICanBoogie\Render\Renderer;
 use ICanBoogie\Render\TemplateResolver;
+use function ICanBoogie\app;
+use function ICanBoogie\get_autoconfig;
 
 class Hooks
 {
@@ -31,7 +31,7 @@ class Hooks
 	 * @param TemplateResolver\AlterEvent $event
 	 * @param TemplateResolver $target
 	 */
-	static public function on_alter_template_resolver(TemplateResolver\AlterEvent $event, TemplateResolver $target)
+	static public function on_alter_template_resolver(TemplateResolver\AlterEvent $event, TemplateResolver $target): void
 	{
 		$event->instance = new ApplicationTemplateResolver($event->instance, get_autoconfig()['app-paths']);
 	}
@@ -40,7 +40,7 @@ class Hooks
 	 * @param EngineCollection\AlterEvent $event
 	 * @param EngineCollection $target
 	 */
-	static public function on_alter_engines(EngineCollection\AlterEvent $event, EngineCollection $target)
+	static public function on_alter_engines(EngineCollection\AlterEvent $event, EngineCollection $target): void
 	{
 		foreach (app()->configs[RenderConfig::DERIVED_ENGINES] as $extension => $engine)
 		{
@@ -57,7 +57,7 @@ class Hooks
 	 *
 	 * @return EngineCollection
 	 */
-	static public function get_template_engines()
+	static public function get_template_engines(): EngineCollection
 	{
 		return Render\get_engines();
 	}
@@ -67,7 +67,7 @@ class Hooks
 	 *
 	 * @return TemplateResolver
 	 */
-	static public function get_template_resolver()
+	static public function get_template_resolver(): TemplateResolver
 	{
 		return clone Render\get_template_resolver();
 	}
@@ -77,7 +77,7 @@ class Hooks
 	 *
 	 * @return Renderer
 	 */
-	static public function get_renderer()
+	static public function get_renderer(): Renderer
 	{
 		return clone Render\get_renderer();
 	}
@@ -89,9 +89,9 @@ class Hooks
 	 * @param mixed $target_or_options
 	 * @param array $additional_options
 	 *
-	 * @return mixed
+	 * @return string|null
 	 */
-	static public function render(Application $app, $target_or_options, array $additional_options = [])
+	static public function render(Application $app, $target_or_options, array $additional_options = []): ?string
 	{
 		return $app->renderer->render($target_or_options, $additional_options);
 	}
