@@ -32,7 +32,7 @@ final class RenderConfig
 	 *
 	 * @return array
 	 */
-	static public function synthesise_engines(array $fragments)
+	static public function synthesise_engines(array $fragments): array
 	{
 		$engines = [];
 
@@ -46,6 +46,6 @@ final class RenderConfig
 			$engines[] = $fragment[self::ENGINES];
 		}
 
-		return $engines ? array_merge(...$engines) : [];
+		return $engines ? \array_merge(...$engines) : [];
 	}
 }
